// client-app/src/app/models/project.ts
export interface IProject {
    id: string;
    createdAt: string;
    updatedAt: string;
    projectCode: string;
    jobType: string;
    orderNumber: string;
    materialOrderNo: string;
    status: string;
    address: string;
    jobStartDate: string;
    estimatedCompletionDate: string;
    startTime: string;
    endTime: string;
    invoiceNo: string;
    remark: string;
}